fix `delete_file` method

+++ mango-orm/src/models/db_query_api/paladins.rs
@@ -74,6 +74,7 @@ pub trait QPaladins: ToModel + CachingModel {
         model_name: &str,
         field_name: &str,
         widget_default_value: &str,
+        is_image: bool,
     ) -> Result<(), Box<dyn std::error::Error>> {
         let hash = self.get_hash().unwrap_or_default();
         if !hash.is_empty() {
@@ -95,14 +96,16 @@ pub trait QPaladins: ToModel + CachingModel {
                             fs::remove_file(path)?;
                         }
                         // Remove thumbnails.
-                        let size_names: [&str; 4] = ["lg", "md", "sm", "xs"];
-                        for size_name in size_names.iter() {
-                            let key_name = format!("{}_{}", "path", size_name);
-                            let path = field_file.get_str(key_name.as_str())?;
-                            if path.is_empty() {
-                                let path = Path::new(path);
-                                if path.exists() {
-                                    fs::remove_file(path)?;
+                        if is_image {
+                            let size_names: [&str; 4] = ["lg", "md", "sm", "xs"];
+                            for size_name in size_names.iter() {
+                                let key_name = format!("{}_{}", "path", size_name);
+                                let path = field_file.get_str(key_name.as_str())?;
+                                if path.is_empty() {
+                                    let path = Path::new(path);
+                                    if path.exists() {
+                                        fs::remove_file(path)?;
+                                    }
                                 }
                             }
                         }
@@ -675,6 +678,7 @@ pub trait QPaladins: ToModel + CachingModel {
                                     model_name,
                                     field_name,
                                     final_widget.value.as_str(),
+                                    false,
                                 )?;
                                 final_doc.insert(field_name, mongodb::bson::Bson::Null);
                             }
@@ -765,6 +769,7 @@ pub trait QPaladins: ToModel + CachingModel {
                                     model_name,
                                     field_name,
                                     final_widget.value.as_str(),
+                                    true,
                                 )?;
                                 final_doc.insert(field_name, mongodb::bson::Bson::Null);
                             }
